pyproject.toml
[project]
name = "pokemon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "pokemontcgsdk>=3.4.0",
    "pytesseract>=0.3.13",
    "requests>=2.32.4",
    "torch>=2.7.1",
    "transformers>=4.52.4",
]
